+++ src/main/java/com/example/eightflix/domain/movie/controller/MovieController.java
@@ -1,8 +1,8 @@
-package com.example.eightflix.domain.movie.Controller;
+package com.example.eightflix.domain.movie.controller;
 
-import com.example.eightflix.domain.movie.MovieDto.MovieRequestDto;
-import com.example.eightflix.domain.movie.MovieDto.MovieResponseDto;
-import com.example.eightflix.domain.movie.Service.MovieService;
+import com.example.eightflix.domain.movie.dto.MovieRequestDto;
+import com.example.eightflix.domain.movie.dto.MovieResponseDto;
+import com.example.eightflix.domain.movie.service.MovieService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
