+++ src/main/java/com/example/eightflix/domain/movie/repository/MovieRepository.java
@@ -1,6 +1,6 @@
-package com.example.eightflix.domain.movie.Repository;
+package com.example.eightflix.domain.movie.repository;
 
-import com.example.eightflix.domain.movie.Entity.Movie;
+import com.example.eightflix.domain.movie.entity.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MovieRepository extends JpaRepository<Movie, Long> {
